refactor to make OpenAI vs Claude request clear
